Order execution refactoring: - Separate persistence models from entities used throughout code - Remove asset data from bundles - Rename BundleData to DataBundle - Move DataBundle to exchange instead of using it directly - Add support for algorthm custom configurations - Add support for multiple exchanges

+++ ziplime/finance/finance_ext.py
@@ -1,9 +1,10 @@
+from decimal import Decimal
 from math import sqrt
 
 import numpy as np
 import pandas as pd
 
-from ziplime.assets.domain.db.futures_contract import FuturesContract
+from ziplime.assets.entities.futures_contract import FuturesContract
 
 
 def update_position_last_sale_prices(positions, get_price, dt):
@@ -13,7 +14,7 @@ def update_position_last_sale_prices(positions, get_price, dt):
     ----------
     positions : OrderedDict
         The positions to update.
-    get_price : callable[Asset, float]
+    get_price : callable[Asset, Decimal]
         The function to retrieve the price for the asset.
     dt : datetime.datetime
         The dt to set as the last sale date if the price is not nan.
@@ -94,14 +95,14 @@ class PositionStats:
     # cdef object underlying_index_array
 
     def __init__(self):
-        self.gross_exposure = 0.0
-        self.gross_value = 0.0
-        self.long_exposure = 0.0
-        self.long_value = 0.0
-        self.net_exposure = 0.0
-        self.net_value = 0.0
-        self.short_exposure = 0.0
-        self.short_value = 0.0
+        self.gross_exposure = Decimal(0.0)
+        self.gross_value = Decimal(0.0)
+        self.long_exposure = Decimal(0.0)
+        self.long_value = Decimal(0.0)
+        self.net_exposure = Decimal(0.0)
+        self.net_value = Decimal(0.0)
+        self.short_exposure = Decimal(0.0)
+        self.short_value = Decimal(0.0)
         self.longs_count = 0
         self.shorts_count = 0
         self.position_exposure_array = None
@@ -144,11 +145,11 @@ def calculate_position_tracker_stats(positions, stats):
         stats.underlying_value_array
     )
 
-    long_value = 0.0
-    short_value = 0.0
+    long_value = Decimal(0.0)
+    short_value = Decimal(0.0)
 
-    long_exposure = 0.0
-    short_exposure = 0.0
+    long_exposure = Decimal(0.0)
+    short_exposure = Decimal(0.0)
 
     longs_count = 0
     shorts_count = 0
